+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.10.1 (unreleased)
 
+- Added `if_not_exists` option to `create_namespace` method
 - Fixed installation error
 
 ## 0.10.0 (2025-08-27)