+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.10.1 (unreleased)
+## 0.10.1 (2025-08-27)
 
 - Added `if_not_exists` option to `create_namespace` method
 - Added `snapshot_id` and `storage_options` options to `to_polars` method